% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{apacorr}
\alias{apacorr}
\title{Correlation Matrix APA}
\usage{
apacorr(X, vars = NULL, digits = 2)
}
\arguments{
\item{X}{a matrix, data.frame, or data table}

\item{vars}{(optional; Default = `NULL`) Character vector on which compute the function.}

\item{digits}{(optional; Default = 2)}
}
\value{
Correlation matrix in APA format (HTML)
}
\description{
Correlation matrix in APA format (HTML)
}
\examples{
apacorr(mtcars)
}
