% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{apatheme}
\alias{apatheme}
\title{APA plot theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
apatheme
}
\value{
Format for ggplot2
}
\description{
Formatted APA plot
}
\examples{
df <- data.frame(x = 1:5, y = 1:5)
ggplot2::ggplot(df, aes(x, y)) + apatheme
}
\keyword{datasets}
