% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ci95h}
\alias{ci95h}
\title{95% high confidence interval}
\usage{
ci95h(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
scalar with high 95%CI
}
\description{
Compute the 95% high confidence interval of the mean
}
\examples{
ci95h(1:10)
}
