% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{extract_num}
\alias{extract_num}
\title{Extract number}
\usage{
extract_num(x)
}
\arguments{
\item{x}{a (vector) string}
}
\value{
Numeric
}
\description{
Extracts number (positive or negative) from a string and converts it to numerical.
If two number in the same string, then it will return the numbers combined
}
\examples{
extract_num("I am 22 years old")
extract_num(c("I am -22 years old", "-74 and 80"))
}
