% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{extractNumber}
\alias{extractNumber}
\title{Extract number from string}
\usage{
extractNumber(x, convertToNum = TRUE)
}
\arguments{
\item{x}{A string.}
}
\value{
Extracted number.
}
\description{
Extracts a number from a string
}
\examples{
extractNumber("A1")
}
