% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genCI}
\alias{genCI}
\title{Generate classification image}
\usage{
genCI(res, stim, N)
}
\arguments{
\item{res}{Numerical vector with the reponses.}

\item{stim}{Numerical vector specifying the stimuli number.}

\item{N}{Matrix of noise pattern as generated with \code{\link[rcicr]{generateStimuli2IFC}}.}
}
\value{
CI
}
\description{
Generate the Classification Image (CI) over a given set of responses.
}
\examples{
NULL
}
