% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genCombinedImg}
\alias{genCombinedImg}
\title{Generate combined image}
\usage{
genCombinedImg(CI, baseImg, filename = NULL, outpath = "./cis",
  size = 512)
}
\arguments{
\item{CI}{Numerical vector of the CI, as generated with \code{\link{genCI}}.}

\item{baseImg}{Numerical vector containing the baseImg image or string pointing to the baseImg image file.
If baseImg is a string, then the baseImg image must in .png or .jpeg.}

\item{outpath}{String specifying the output target path}
}
\description{
Generate the combinaed image by imposing the selected noise on the base image.
}
\examples{
NULL
}
