% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genMask}
\alias{genMask}
\title{Generate (and scale) mask from responses}
\usage{
genMask(response, stim, noiseMatrix, baseImg, scaling = "matched")
}
\arguments{
\item{response}{Numerical vector specifying the reponses.}

\item{stim}{Numerical vector specifying the stimuli number.}

\item{noiseMatrix}{Matrix of noise pattern as generated with
\code{noiseMatrix <- rcirc::generateStimuli2IFC(..., return_as_dataframe = TRUE)}.}

\item{baseImg}{Numerical vector containing the baseImg image or string pointing to the baseImg
image file. If baseImg is a string, then the baseImg image must in .png or .jpeg.}

\item{scaling}{String|Scalar|NULL specifying the scaling method. `"matched"` is the default method.
If a scalar is provided (e.g. 5) than the `"constant"` method will be applied.
If `NULL` no scaling is applied.}
}
\value{
List with the (un)scaled Noise mask (\code{$mask}) and the base image as a vector
  (\code{$baseImgVect}).
}
\description{
Generate (and scale) mask from responses.
}
\examples{
NULL
}
