% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genNM}
\alias{genNM}
\title{Generate (scaled) noise mask from responses}
\usage{
genNM(res, stim, N, baseImg, scaling = "matched")
}
\arguments{
\item{res}{Numerical vector with the reponses.}

\item{stim}{Numerical vector specifying the stimuli number.}

\item{N}{Matrix of noise pattern as generated with
\code{\link[rcicr]{N <- generateStimuli2IFC(..., return_as_dataframe = TRUE)}}.}

\item{baseImg}{Numerical vector containing the baseImg image or string pointing to the baseImg image file.
If baseImg is a string, then the baseImg image must in .png or .jpeg.}

\item{scaling}{String|Scalar|NULL specifying the scaling method. `"matched"` is the default method.
If a scalar is provided (e.g. 5) than the `"constant"` method will be applied.
If `NULL` no scaling is applied.}
}
\value{
(Un)scaled Noise mask (NM).
}
\description{
Generate the noise mask (NM) over a given set of responses.
}
\examples{
NULL
}
