% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{getFaceRegion}
\alias{getFaceRegion}
\title{Get face region}
\usage{
getFaceRegion(imgPath, xpos = 0, ypos = 0, faceWidth = 1.4,
  faceHeight = 1.8, preview = TRUE, writeImgTo = NULL)
}
\arguments{
\item{imgPath}{String specifying path to the base image.}

\item{xpos}{Numeric specifiying the X position (relative to the center).}

\item{ypos}{Numeric specifiying the Y position (relative to the center).}

\item{faceWidth}{Numeric specifiying the width of the face region.}

\item{faceHeight}{Numeric specifiying the height of the face region.}

\item{preview}{Numeric specifiying the height of the face region.}

\item{writeImgTo}{String specifying where and if the output image should be saved. Default is
NULL, meaning that the image will not be saved.}
}
\value{
Logical vector specifying the location of the face region
}
\description{
Returns a logical vector with the face region
}
\examples{
NULL
}
