% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Value Unmatching}
\usage{
... \%ni\% NA
}
\arguments{
\item{x}{A string.}
}
\value{
Extracted number.
}
\description{
Inverse of \code{\link[base]{match}}.
}
\examples{
1 \%ni\% (1:3)
}
