% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWrangling.R
\name{loj}
\alias{loj}
\title{Schmitz Package}
\usage{
loj(X = NULL, Y = NULL, onCol = NULL)
}
\arguments{
\item{X}{a data.table}

\item{Y}{a data.table}

\item{onCol}{a string or vector common to X and Y on which to merge the data}
}
\description{
Left outter join on X
}
\examples{
NULL
}
