% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{m_sd}
\alias{m_sd}
\title{Formatted mean and standard deviation}
\usage{
m_sd(x, ...)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
formatted string
}
\description{
Format the mean and sd of a vector
}
\examples{
m_sd(rnorm(100))
}
