% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{mat2img}
\alias{mat2img}
\title{Matrix to image}
\usage{
mat2img(img, density = NULL, depth = NULL, strip = FALSE, alpha = 1)
}
\arguments{
\item{img}{Image (e.g., image path) to be read, or numerical object (e.g., matrix, array)
see \code{magick::image_read()} for more information (equivalent to `path` argument).}

\item{density}{see \code{magick::image_read()}}

\item{depth}{see \code{magick::image_read()}}

\item{strip}{see \code{magick::image_read()}}

\item{alpha}{(optional) Scalar indicating the trasnparency of the (alpha). If the object already
contains an alpha, than it will not be overwritten. Default is `1`.}
}
\description{
Generalized version of \code{magick::image_read()} which also
 allows to read numerical objects.
}
\examples{
NULL
}
