% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\docType{data}
\name{N}
\alias{N}
\title{noiseMatrix data}
\format{A (512^2)x20 matrix of noise as generated with
  \code{N <- matrix(rcicr::generateCI2IFC(..., return_as_dataframe = TRUE))}.}
\usage{
data(N)
}
\description{
Fake noiseMatrix data
}
\keyword{datasets}
