% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{numerify}
\alias{numerify}
\title{Numerify}
\usage{
numerify(X)
}
\arguments{
\item{X}{Data.table}
}
\description{
Tries to convert each column of a data.table `X` to numeric (except logicals).
}
\examples{
NULL
}
