% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pprint}
\alias{pprint}
\title{Print p-values}
\usage{
pprint(pval)
}
\arguments{
\item{mdl}{an lm model obtained from lm}
}
\value{
data table with summary statistics
}
\description{
Nicely print p-values (*p<.05, **p<.01, ***p<.001)
}
\examples{
pprint(.003)
}
