% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{quiet}
\alias{quiet}
\title{Supress output}
\usage{
quiet(x)
}
\arguments{
\item{x}{a function with output}
}
\value{
The function output
}
\description{
Completely silence the output from a function
}
