% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\docType{data}
\name{RC}
\alias{RC}
\title{RC data}
\format{A data frame with 80 rows and 4 variables:
\describe{
  \item{id}{Participant's unique identifier.}
  \item{stimuli}{Stimuli number. This is pared with the noiseMatrix (\code{N}), e.g. stimuli 1
    was generated from the base image and the first column of N (\code{N[, 1]}).}
  \item{reponses}{Responses from participant. Can be \code{1} (oriented image selected) or
    \code{-1} (inverted image selected).}
  \item{condition}{Two conditions: \code{A} and \code{B}.}
}}
\usage{
data(RC)
}
\description{
Fake RC data
}
\keyword{datasets}
