% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{round2}
\alias{round2}
\title{Round 2}
\usage{
round2(x)
}
\arguments{
\item{x}{a number or a numeric vector}
}
\value{
a string or a character vector
}
\description{
Rounds the number to two decimal points.
If the value < .01 than a label "<.01" is used.
Removes leading zeros.
}
\examples{
round2(c(0.033, 0.1, 0.0002))
}
