% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{roundify}
\alias{roundify}
\title{Roundify}
\usage{
roundify(X, digits)
}
\arguments{
\item{X}{Data.table}
}
\description{
Try to counrd each numerical column of a data.table.
}
\examples{
NULL
}
