% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{scaleCI}
\alias{scaleCI}
\title{Scale Classification Image}
\usage{
scaleCI(CI, baseImg, scaling = "matched")
}
\arguments{
\item{CI}{Numerical vector of the CI, as generated with \code{\link{genCI}}.}

\item{baseImg}{Numerical vector containing the baseImg image or string pointing to the baseImg image file.
If baseImg is a string, then the baseImg image must in .png or .jpeg.}

\item{scaling}{String|Scalar|NULL specifying the scaling method. `"matched"` is the default method.
If a scalar is provided (e.g. 5) than the `"constant"` method will be applied.
If `NULL` no scaling is applied.}
}
\value{
Scaled CI.
}
\description{
Scales the Cassification Image (CI).
}
\examples{
NULL
}
