% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmitz.R
\docType{package}
\name{schmitz}
\alias{schmitz}
\title{schmitz: My first package, just some custom functions.}
\description{
The \code{schmitz} package provides three categories of functions:
RC, stats, and dataWranglingCleaning.
}
\section{RC functions}{

The RC functions ...
}

\section{stats functions}{

The stats functions ..
}

\section{dataWranglingCleaning functions}{

The dataWranglingCleaning functions ..
}

