% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
scalar
}
\description{
Compute the stardard error of the mean (SE)
}
\examples{
sem(1:10)
}
