% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ss}
\alias{ss}
\title{LMM and LM results summary}
\usage{
ss(mdl)
}
\arguments{
\item{mdl}{an lm model obtained from lm}
}
\value{
data table with summary statistics
}
\description{
Formats the results from LMM and LM. Provides effect sizes.
}
\examples{
ss(mdl)
}
