% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ss_anova}
\alias{ss_anova}
\title{Anova results summary}
\usage{
ss_anova(mdl)
}
\arguments{
\item{mdl}{an lm(er) model obtained from lm(er)}
}
\value{
data table with summary statistics
}
\description{
Formats the results from Anova and provides effect sizes.
}
\examples{
ss_anova(mdl)
}
