% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ssBF}
\alias{ssBF}
\title{Report Bayes factor}
\usage{
ssBF(BFobj, digits = 2)
}
\arguments{
\item{BFobj}{a `BFBayesFactor` object as returned from `lmBF(mdlH1)/lmBF(mdlH0)`. Warning, 
the `BFobj` must be in favor of the alternative hypothesis (BF10)}

\item{digits}{number of decimal digits}
}
\value{
String with the BF10, BF01 and error
}
\description{
Report the Bayes factor (BF10, BF01 and error)
}
