% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{apacorr}
\alias{apacorr}
\title{Correlation Matrix APA}
\usage{
apacorr(
  X,
  vars = NULL,
  w = NULL,
  digits = 2,
  kableFormat = TRUE,
  kableColor = TRUE
)
}
\arguments{
\item{X}{a matrix, data.frame, or data table}

\item{vars}{(optional; Default = `NULL`) Character vector on which compute the function.}

\item{w}{(optional; Default = `NULL`) Weight vector to be applied to the analyses.}

\item{digits}{(optional; Default = 2)}

\item{kableFormat}{(optional; Default = `TRUE`) Should the table be formated with kableExtra}

\item{kableColor}{(optional; Default = `TRUE`) Color formatble table (only if `kableFormat = TRUE`)}
}
\value{
Correlation matrix in APA format (in HTML if `kableExtra = TRUE`).
}
\description{
Correlation matrix in APA format
}
\examples{
apacorr(mtcars)
}
