% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ci95}
\alias{ci95}
\title{95% confidence interval}
\usage{
ci95(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
vector of length 2: the first value is the low 95%CI and the second the high 95%ci
}
\description{
Compute the 95% confidence interval of the mean
}
\examples{
ci95(1:10)
}
