% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ci95}
\alias{ci95}
\title{95 percent confidence interval}
\usage{
ci95(x)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
vector of length 2: the first value is the low 95 percent confidence interval
 and the second the high 95 percent confidence interval
}
\description{
Compute the 95 percent confidence interval of the mean
}
\examples{
ci95(1:10)
}
