% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genCI}
\alias{genCI}
\title{Generate Classification Image (CI)}
\usage{
genCI(NM = NULL, filename = "combined.png", outpath = "./cis",
  resize = NULL, preview = FALSE, ...)
}
\arguments{
\item{NM}{Noise mask as generated from \code{\link{genNM}}.}

\item{filename}{String specifying the name of the ouput CI. Default is `"combined.png"`.}

\item{outpath}{String specifying the output target path.}

\item{resize}{Scalar specifying if the image should be resized.}

\item{preview}{Logical indicating if the iamge should be previewed in the Viewer Panel.}

\item{...}{List of parameters passed to \code{\link{genNM}}. If `NM` is provided, then ellipsis
will be ignored.}
}
\description{
Generate the combinaed of the noise mask (NM) and the base image.
}
\examples{
NULL
}
