% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genCI}
\alias{genCI}
\title{Generate Classification Image (CI)}
\usage{
genCI(response, stim, noiseMatrix, baseImg, scaling = "matched",
  outputPath = "combined.png")
}
\arguments{
\item{response}{Numerical vector specifying the reponses.}

\item{stim}{Numerical vector specifying the stimuli number.}

\item{noiseMatrix}{Matrix of noise pattern as generated with
\code{\link[rcicr]{noiseMatrix <- generateStimuli2IFC(..., return_as_dataframe = TRUE)}}.}

\item{baseImg}{Numerical vector containing the baseImg image or string pointing to the baseImg
image file. If baseImg is a string, then the baseImg image must in .png or .jpeg.}

\item{scaling}{String|Scalar|NULL specifying the scaling method. `"matched"` is the default method.
If a scalar is provided (e.g. 5) than the `"constant"` method will be applied.
If `NULL` no scaling is applied.}

\item{outputPath}{String specifying the file path of the ouput CI. Default is `"combined.png"`.}
}
\description{
Generate the combinaed of the noise mask (mask) and the base image.
}
\examples{
NULL
}
