% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{genCI}
\alias{genCI}
\title{Generate Classification Image (CI)}
\usage{
genCI(mask = NULL, filename = "combined.png",
  outpath = "./cis/combined.png", resize = NULL, preview = FALSE,
  ...)
}
\arguments{
\item{mask}{Mask as generated from \code{\link{genMask}}. If `NULL` (default), then the `mask`
will be generated automatically by calling \code{\link{genMask}}. See `...`.}

\item{filename}{String specifying the name of the ouput CI. Default is `"combined.png"`.}

\item{outpath}{String specifying the output target path.}

\item{resize}{Scalar specifying if the image should be resized.}

\item{preview}{Logical indicating if the image should be previewed in the Viewer Panel.}

\item{...}{List of parameters passed to \code{\link{genMask}}. If `mask` is provided, then ellipsis
will be ignored.}
}
\description{
Generate the combinaed of the noise mask (mask) and the base image.
}
\examples{
NULL
}
