% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loj.R
\name{loj}
\alias{loj}
\title{Left Outter Join}
\usage{
loj(X = NULL, Y = NULL, onCol = NULL)
}
\arguments{
\item{X, Y}{a data.tables}

\item{onCol}{a string or character vector common to `X` and `Y` on which to merge the data}
}
\description{
Update and merge columuns from `Y` onto `X`. If data tables are loaded, make sure
 to run `setDT(X)` and `setDT(Y)`.
}
\examples{
NULL
}
