% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sd.R
\name{m_sd}
\alias{m_sd}
\title{Formated mean and standard deviation}
\usage{
m_sd(x, ...)
}
\arguments{
\item{x}{a numerical vector}
}
\value{
formated string
}
\description{
Format the mean and sd of a vector
}
\examples{
m_sd(1:3)
}
