% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.R
\name{readBaseImg}
\alias{readBaseImg}
\title{Read and scale base image}
\usage{
readBaseImg(baseImgPath, maxContrast = TRUE)
}
\arguments{
\item{baseImgPath}{String specifying path to the base image.}
}
\value{
Vector of pixels from the base image.
}
\description{
Read and scale the base image
}
\examples{
NULL
}
