% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataWranglingCleaning.R
\name{roundify}
\alias{roundify}
\title{Roundify}
\usage{
roundify(X, digits = 2)
}
\arguments{
\item{X}{data.table}

\item{digits}{integer}

\item{printOut}{logical that indicates if the output needs to be printed}
}
\description{
Try to round each numerical column of a data.table and the prints it. Usefull with
  rmarkdown reports.
}
\examples{
NULL
}
