% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ss}
\alias{ss}
\title{Effect sizes for lm summary}
\usage{
ss(mdl, digits = 3, printOut = TRUE)
}
\arguments{
\item{mdl}{an lm model obtained from lm}

\item{digits}{number of decimal digits}

\item{printOut}{prints the output}
}
\value{
data table with summary statistics
}
\description{
Add effect sizes to the lm summary
}
\examples{
ss(mdl)
}
