% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ssBF}
\alias{ssBF}
\title{LMM and LM results summary with Bayes factor}
\usage{
ssBF(mdl)
}
\arguments{
\item{mdl}{an lm model obtained from lm}
}
\value{
data table with summary statistics
}
\description{
Formats the results from LMM and LM. Provides effect sizes and Bayes factor.
}
\examples{
ssb(mdl)
}
